+++ cardano/src/block/verify.rs
@@ -33,7 +33,7 @@ pub enum Error {
     WrongBlockHash,
     WrongDelegationProof,
     WrongExtraDataProof,
-    WrongGenesisProof,
+    WrongBoundaryProof,
     WrongMagic,
     WrongMerkleRoot,
     WrongMpcProof,
@@ -79,7 +79,7 @@ impl fmt::Display for Error {
             WrongBlockHash => write!(f, "block hash is invalid"),
             WrongDelegationProof => write!(f, "delegation proof is invalid"),
             WrongExtraDataProof => write!(f, "extra data proof is invalid"),
-            WrongGenesisProof => write!(f, "genesis proof is invalid"),
+            WrongBoundaryProof => write!(f, "boundary proof is invalid"),
             WrongMagic => write!(f, "magic number is invalid"),
             WrongMerkleRoot => write!(f, "merkle root is invalid"),
             WrongMpcProof => write!(f, "MPC proof is invalid"),
@@ -143,7 +143,7 @@ pub fn verify_block(protocol_magic: ProtocolMagic,
     Ok(())
 }
 
-impl Verify for genesis::Block {
+impl Verify for boundary::Block {
     fn verify(&self, protocol_magic: ProtocolMagic) -> Result<(), Error> {
         let hdr = &self.header;
 
@@ -153,7 +153,7 @@ impl Verify for genesis::Block {
 
         // check body proof
         if hash::Blake2b256::new(&cbor!(&self.body).unwrap()) != hdr.body_proof.0 {
-            return Err(Error::WrongGenesisProof);
+            return Err(Error::WrongBoundaryProof);
         }
 
         Ok(())
